+++ benchmarks/bench_ascon_xof.cpp
@@ -1,5 +1,5 @@
+#include "ascon/hashing/ascon_xof.hpp"
 #include "bench_helper.hpp"
-#include "hashing/ascon_xof.hpp"
 #include <benchmark/benchmark.h>
 #include <span>
 #include <vector>
