+++ include/ascon/auth/ascon_mac.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "ascon_perm.hpp"
+#include "ascon/ascon_perm.hpp"
 #include "sponge.hpp"
 
 // Ascon-MAC ( message authentication code )
