+++ include/ascon/hashing/ascon_hasha.hpp
@@ -1,9 +1,9 @@
 #pragma once
-#include "ascon_perm.hpp"
+#include "ascon/ascon_perm.hpp"
 #include "sponge.hpp"
 
-// Ascon-XofA
-namespace ascon_xofa {
+// Ascon-Hasha
+namespace ascon_hasha {
 
 // Bit width of rate portion of Ascon permutation state
 constexpr size_t RATE = 64;
@@ -14,30 +14,33 @@ constexpr size_t ROUNDS_A = 12;
 // How many rounds of Ascon permutation is applied for p^b
 constexpr size_t ROUNDS_B = 8;
 
-// 64 -bit initialization value, obtained from section 2.5.1 of spec.
-constexpr uint64_t IV = 0x00400c0400000000ul;
+// Ascon HashA Digest Byte Length
+constexpr size_t DIGEST_LEN = sizeof(ascon_perm::ascon_perm_t) - (RATE / 8);
 
-// Compile-time computed initial permutation state used in Ascon-XofA.
+// 64 -bit initialization value, obtained from section 2.5.1 of spec.
+constexpr uint64_t IV = 0x00400c0400000100ul;
+
+// Compile-time computed initial permutation state used in Ascon-HashA.
 constexpr auto INIT_PERM_STATE = sponge::compute_init_state(IV);
 
-// Ascon-XofA ( extendable output function ), supporting incremental hashing.
+// Ascon-HashA, supporting incremental hashing.
 //
 // See section 2.5 of Ascon specification
 // https://ascon.iaik.tugraz.at/files/asconv12-nist.pdf
-struct ascon_xofa_t
+struct ascon_hasha_t
 {
 private:
   ascon_perm::ascon_perm_t state = INIT_PERM_STATE;
   size_t offset = 0;
-  size_t readable = 0;
   alignas(4) bool absorbed = false;
+  alignas(4) bool squeezed = false;
 
 public:
   // Constructor
-  inline constexpr ascon_xofa_t() = default;
+  inline constexpr ascon_hasha_t() = default;
 
-  // Given N -bytes input message, this routine consumes those into Ascon
-  // permutation state.
+  // Given N -bytes input message, this routine consumes those into
+  // Ascon permutation state.
   //
   // Note, this routine can be called arbitrary number of times, each time with
   // arbitrary bytes of input message, until Ascon permutation state is
@@ -54,48 +57,41 @@ public:
   // no more input bytes remaining to be consumed by Ascon permutation state.
   //
   // Note, once this routine is called, calling absorb() or finalize() again, on
-  // same Ascon-XOFA object, doesn't do anything. After finalization, one would
-  // like to read arbitrary many bytes of digest by squeezing sponge, which is
-  // done by calling squeeze() function as many times required.
+  // same Ascon-HashA object, doesn't do anything. After finalization, one would
+  // like to read 32 -bytes of digest by squeezing sponge, which is done by
+  // calling digest() function only once.
   inline constexpr void finalize()
   {
     if (!absorbed) {
       sponge::finalize<ROUNDS_A, RATE>(state, offset);
-
       absorbed = true;
-      readable = RATE / 8;
     }
   }
 
-  // Given that N -bytes input message is already absorbed into Ascon permutation state,
-  // this routine is used for squeezing M -bytes out of consumable part of state ( i.e.
-  // rate portion of state ).
-  //
-  // This routine can be used for squeezing arbitrary number of bytes from Ascon
-  // permutation state.
-  //
-  // Make sure you absorb message bytes first, then only call this function, otherwise
-  // it can't squeeze out anything. Once done with squeezing, you can reuse same hasher
-  // object for another absorb->finalize->squeeze round, by explicitly calling reset()
-  // function.
-  inline constexpr void squeeze(std::span<uint8_t> out)
+  // Given that N -bytes message is consumed into Ascon permutation state, this routine
+  // can be used for squeezing out 32 -bytes message digest. Once squeezed, calling this
+  // function again and again doesn't have any effect. You may continue using same
+  // hasher object for another round of absorb->finalize->digest flow, by explicitly
+  // calling reset() function.
+  inline constexpr void digest(std::span<uint8_t, DIGEST_LEN> out)
   {
-    if (!absorbed) {
-      return;
-    }
+    if (absorbed && !squeezed) {
+      size_t readable = RATE / 8;
+      sponge::squeeze<ROUNDS_B, RATE>(state, readable, out);
 
-    sponge::squeeze<ROUNDS_B, RATE>(state, readable, out);
+      squeezed = true;
+    }
   }
 
-  // Resets internal state of Ascon-XofA object, making it ready for yet another
-  // absorb->finalize->squeeze round. You must explicitly call this function if you're
+  // Resets internal state of Ascon-HashA object, making it ready for yet another
+  // absorb->finalize->digest round. You must explicitly call this function if you're
   // interested in reusing the same object for absorbing new messages.
   inline constexpr void reset()
   {
     this->state = INIT_PERM_STATE;
     this->offset = 0;
-    this->readable = 0;
     this->absorbed = false;
+    this->squeezed = false;
   }
 };
