+++ include/ascon/utils.hpp
